re-factoring the content manager class a bit
