+++ src/Microweber/Application.php
@@ -40,6 +40,7 @@ namespace Microweber;
  * @property \Microweber\Providers\FormsManager              $forms_manager
  * @property \Microweber\Providers\Content\DataFieldsManager $data_fields_manager
  * @property \Microweber\Providers\Content\AttributesManager $attributes_manager
+ * @property \Microweber\Providers\Content\ContentManagerCrud       $content_manager_crud
  *
  */
 class Application {
