+++ src/Microweber/MicroweberServiceProvider.php
@@ -63,13 +63,7 @@ class MicroweberServiceProvider extends ServiceProvider {
             return new Providers\DatabaseManager($app);
         });
 
-        $this->app->singleton('crud', function ($app) {
-            return new Providers\Database\Crud($app);
-        });
 
-        $this->app->singleton('user_crud', function ($app) {
-            return new Providers\Database\UserCrud($app);
-        });
         $this->app->singleton('format', function ($app) {
             return new Utils\Format($app);
         });
@@ -109,6 +103,7 @@ class MicroweberServiceProvider extends ServiceProvider {
         $this->app->singleton('fields_manager', function ($app) {
             return new Providers\FieldsManager($app);
         });
+
 
         $this->app->singleton('data_fields_manager', function ($app) {
             return new Providers\Content\DataFieldsManager($app);
@@ -188,6 +183,13 @@ class MicroweberServiceProvider extends ServiceProvider {
             return new Providers\Ui($app);
         });
 
+        $this->app->singleton('content_manager_crud', function ($app) {
+            return new Providers\Content\ContentManagerCrud($app);
+        });
+
+        $this->app->singleton('content_manager_helpers', function ($app) {
+            return new Providers\Content\ContentManagerHelpers($app);
+        });
 
         $this->app->register('Illuminate\Html\HtmlServiceProvider');
         AliasLoader::getInstance()->alias("Form", 'Illuminate\Html\FormFacade');
