+++ src/Microweber/Providers/Content/AttributesManager.php
@@ -104,7 +104,8 @@ class AttributesManager extends Crud {
             }
         }
         $save = parent::save($data);
-         return $save;
+
+        return $save;
     }
